Bugfixes and Deli example
Still a bit worried there are some issues here, but this fixed a bug
with sleeping and clock advancement. The issue appeared to be using
`callC` instead of `shiftT`, and then a missing `dequeue` call. The
missing `dequeue` call is what is concerning, as I don't have a great
intuition about where they all need to be.